+++ src/Foundation/Support/helpers.php
@@ -2,6 +2,8 @@
 
 declare(strict_types=1);
 
+use Vanilo\MasterProduct\Contracts\MasterProduct;
+
 /**
  * Returns the price formatted
  *
@@ -16,4 +18,9 @@ function format_price($price, string $currency = null)
         $price,
         $currency ?? config('vanilo.foundation.currency.sign')
     );
+}
+
+function is_master_product(object $product): bool
+{
+    return $product instanceof MasterProduct;
 }